Revert "Made crafting obey remaining items, and added some non-working recipes."
