--- src/main/resources/config.yml
@@ -1,5 +1,0 @@
-craftCooldown: 8
-redstoneMode: 'indirect'
-particles: true
-dropItemsIfNoOutputContainer: true
-craftSounds: true
